Trekk validering ut i egen fil
